+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 ### bug fixes
 
 - don't break if you choose an already taken name for your Heroku app
+- remove the `root_url` from application.html.haml since there's no default `root_url`
 
 ## 0.0.6 (August 16, 2013)